+++ ProgramFunctions/Debug.fnc.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * Debug RosarioSIS
+ */
+
+/**
+ * Load PHP Debug bar.
+ *
+ * @see https://gitlab.com/francoisjacquet/rosariosis-meta#php-debug-bar
+ *
+ * @link https://github.com/maximebf/php-debugbar
+ *
+ * @since 5.0
+ */
+function PhpDebugBar()
+{
+	global $debugbarRenderer;
+
+	if ( ! file_exists( 'meta/debug/php-debugbar/vendor/autoload.php' ) )
+	{
+		return false;
+	}
+
+	// Load PHP debug bar.
+	require_once 'meta/debug/php-debugbar/vendor/autoload.php';
+
+	$debugbar = new DebugBar\StandardDebugBar();
+	$debugbarRenderer = $debugbar->getJavascriptRenderer(
+		'meta/debug/php-debugbar/src/DebugBar/Resources'
+	);
+
+	// Fix $ not defined JS error.
+	$debugbarRenderer->setIncludeVendors('css');
+
+	function debugBarRenderHead()
+	{
+		global $debugbarRenderer;
+
+		echo $debugbarRenderer->renderHead();
+	}
+
+	add_action( 'Warehouse.php|header_head', 'debugBarRenderHead' );
+
+	function debugbarRender()
+	{
+		global $debugbarRenderer;
+
+		echo $debugbarRenderer->render();
+	}
+
+	add_action( 'Warehouse.php|footer', 'debugbarRender' );
+
+	return true;
+}